adk_agent/requirements.txt
google-adk
google-cloud-bigquery
google-auth
pandas
python-dotenv
